refactor: fetch join 사용해서 이모지 조회

+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/expense/service/dto/response/EmojiResponse.java
@@ -1,4 +1,4 @@
 package com.jjangiji.hankkimoa.expense.service.dto.response;
 
-public record EmojiResponse(Integer emojiId, Integer count) {
+public record EmojiResponse(Integer emojiId, Long count) {
 }
